ahora el módulo música se puede activar o desactivar

+++ musica.py
@@ -1,21 +1,57 @@
 import os
 import time
 
-def playMP3 (str_song):
-	os.system("mpg123 --loop -1 --quiet music/" + str(str_song) + ".mp3 &")
-	time.sleep(0.5)
+def initMusica (bool_activar = True):
+	global bool_musicaActivada
+	bool_musicaActivada = bool_activar
 
-def playWAV (str_sound):
-	os.system("afplay music/" + str(str_sound) + ".wav &")
+
+def desactivarMusica ():
+	global bool_musicaActivada
+	bool_musicaActivada = False
+
+
+def playMP3 (str_song, str_directory = "music"):
+	global bool_musicaActivada
+	try:
+		if (bool_musicaActivada):
+			os.system("mpg123 --loop -1 --quiet "+str(str_directory)+"/"+str(str_song)+".mp3 &")
+			time.sleep(0.5)
+	except Exception as e:
+		pass
+
+		
+
+def playWAV (str_sound, str_directory = "music"):
+	global bool_musicaActivada
+	try:
+		if (bool_musicaActivada):
+			os.system("afplay "+str(str_directory)+"/"+str(str_sound)+".wav &")
+	except Exception as e:
+		pass
+	
 
 
 def stopMP3 ():
-	os.system("pkill mpg123")
+	global bool_musicaActivada
+	try:
+		if (bool_musicaActivada):
+			os.system("pkill mpg123")
+	except Exception as e:
+		pass
+	
+
 
 def stopWAV():
-	os.system("pkill afplay")
-
+	global bool_musicaActivada
+	try:
+		if (bool_musicaActivada):
+			os.system("pkill afplay")
+	except Exception as e:
+		pass
 	
+
+
 def stop():
 	stopMP3()
 	stopWAV()
